Fix ban and mute button, add blocked users management

+++ backend/src/routes/chat/chat.service.ts
@@ -68,6 +68,12 @@ export class ChatService {
       if (!dec)
         throw new HttpException("Wrong password", HttpStatus.UNAUTHORIZED);
     }
+
+    if (await this.channelMemberService.getMember(channel.id, id))
+      throw new HttpException(
+        "You already joined this channel",
+        HttpStatus.BAD_REQUEST
+      );
     await this.channelMemberService.addMember(channel.id, id);
 
     return true;
@@ -78,6 +84,12 @@ export class ChatService {
 
     if (!channel)
       throw new HttpException("Channel not found", HttpStatus.BAD_REQUEST);
+    if (channel.ownerId === id)
+      throw new HttpException(
+        "You cannot leave a channel you own.",
+        HttpStatus.BAD_REQUEST
+      );
+
     await this.channelMemberService.removeMember(channel.id, id);
 
     return true;
@@ -225,9 +237,6 @@ export class ChatService {
   addAdmin = async (oid: number, chan: string, userName: string) => {
     const channel = await this.channelService.getChannelByName(chan);
     const user = await this.userService.getUserByLogin(userName);
-    if (await this.channelAdminService.getAdmin(channel?.id, oid)) {
-      throw new HttpException("User is already admin", HttpStatus.BAD_REQUEST);
-    }
     if (!channel || !user)
       throw new HttpException(
         "Channel or user not found",
@@ -235,10 +244,17 @@ export class ChatService {
       );
 
     if (oid === channel.ownerId) {
+      if (await this.channelAdminService.getAdmin(channel?.id, user.id)) {
+        throw new HttpException(
+          "User is already admin",
+          HttpStatus.BAD_REQUEST
+        );
+      }
+
       await this.channelAdminService.addAdmin(channel.id, user.id);
     } else
       throw new HttpException(
-        "Member can only be banned by an admin",
+        "Member can only be made admin by the owner",
         HttpStatus.FORBIDDEN
       );
   };
